organize / graceful shutdown / dynamic templating
