+++ src/components/blocks/hero.go.html
@@ -48,7 +48,7 @@
 
         <div class="mt-10 flex items-center justify-center gap-x-6">
           {{ range .buttons }}
-          <a preload="mouseover" href="{{ .href }}" class="btn-{{ .variant }}">{{ .label }}</a>
+          <a preload="mouseover" href="{{ .href }}" class="btn-{{ .variant }} btn-lg">{{ .label }}</a>
           {{ end }}
         </div>
       </div>
