--- src/components/films.go.html
@@ -1,15 +1,0 @@
-{{ define "films" }}
-<h2>Films</h2>
-<div class="flex flex-wrap">
-  {{ range .Films }}
-  <div class="mb-2 w-full md:w-1/2 lg:w-1/3 p-2">
-    <div class="rounded shadow p-5 border">
-      <h4>{{ .Title }}</h4>
-      <p>Director: {{ .Director }}</p>
-      <p>Year: {{ .Year }}</p>
-      <p>Actor: {{ .Actor }}</p>
-    </div>
-  </div>
-  {{ end }}
-</div>
-{{ end }}
