+++ src/components/head.go.html
@@ -7,28 +7,14 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <link rel="stylesheet" href="/css/main.css?v={{ .Version }}" />
   <link rel="icon" type="image/x-icon" href="/favicon.ico?v={{ .Version }}" />
-  <script
-    defer
-    src="https://unpkg.com/htmx.org@1.9.7/dist/htmx.min.js"
-  ></script>
-  <script
-    defer
-    src="https://unpkg.com/htmx.org@1.9.7/dist/ext/preload.js"
-  ></script>
-
-  <!-- <script defer src="https://unpkg.com/htmx.org/dist/ext/preload.js"></script> -->
+  <script defer src="https://unpkg.com/htmx.org@1.9.7/dist/htmx.min.js"></script>
+  <script defer src="https://unpkg.com/htmx.org@1.9.7/dist/ext/preload.js"></script>
 
   <!-- deferred loading of google font css -->
-  <!-- <link
-    rel="preload"
-    href="https://fonts.googleapis.com/css2?family=Inter&display=swap"
-    as="style"
-    onload="this.onload=null;this.rel='stylesheet'"
-  />
-  <noscript
-    ><link
-      rel="stylesheet"
-      href="https://fonts.googleapis.com/css2?family=Inter&display=swap"
-  /></noscript> -->
+  <link rel="preload" href="https://fonts.googleapis.com/css2?family=Inter&display=swap" as="style"
+    onload="this.onload=null;this.rel='stylesheet'" />
+  <noscript>
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter&display=swap" />
+  </noscript>
 </head>
 {{ end }}
