+++ src/templates/404.go.html
@@ -1,34 +1,27 @@
-<!DOCTYPE html>
-<html lang="en" class="h-full">
-{{ template "head" . }}
+{{ define "content" }}
 
-<body class="flex flex-col justify-between h-screen">
-  {{ template "navbar" . }}
+<div class="grid grid-cols-1 grid-rows-[1fr,auto,1fr] bg-white lg:grid-cols-[max(50%,36rem),1fr] h-full">
+  <main class="mx-auto w-full max-w-7xl px-6 py-24 sm:py-32 lg:col-span-2 lg:col-start-1 lg:row-start-2 lg:px-8">
+    <div class="max-w-lg">
+      <p class="text-base font-semibold leading-8 text-indigo-600">404</p>
+      <h1 class="mt-4 text-3xl font-bold tracking-tight text-gray-900 sm:text-5xl">
+        Woops, you shouldn't be here...
+      </h1>
+      <p class="mt-6 text-base leading-7 text-gray-600">
+        Our server cat got curious and played with the wires. Guess this
+        page was the toy!
+      </p>
 
-  <div class="grid grid-cols-1 grid-rows-[1fr,auto,1fr] bg-white lg:grid-cols-[max(50%,36rem),1fr] h-full">
-    <main class="mx-auto w-full max-w-7xl px-6 py-24 sm:py-32 lg:col-span-2 lg:col-start-1 lg:row-start-2 lg:px-8">
-      <div class="max-w-lg">
-        <p class="text-base font-semibold leading-8 text-indigo-600">404</p>
-        <h1 class="mt-4 text-3xl font-bold tracking-tight text-gray-900 sm:text-5xl">
-          Woops, you shouldn't be here...
-        </h1>
-        <p class="mt-6 text-base leading-7 text-gray-600">
-          Our server cat got curious and played with the wires. Guess this
-          page was the toy!
-        </p>
-
-        <div class="mt-10">
-          <a href="/" class="text-sm font-semibold leading-7 text-indigo-600"><span aria-hidden="true">&larr;</span>
-            Back to home</a>
-        </div>
+      <div class="mt-10">
+        <a href="/" class="text-sm font-semibold leading-7 text-indigo-600"><span aria-hidden="true">&larr;</span>
+          Back to home</a>
       </div>
-    </main>
-
-    <div class="hidden lg:relative lg:col-start-2 lg:row-start-1 lg:row-end-4 lg:block h-full">
-      <img src="https://cataas.com/cat" alt="" class="absolute inset-0 h-full w-full object-cover" />
     </div>
-  </div>
-  {{ template "footer" . }}
-</body>
+  </main>
 
-</html>
+  <div class="hidden lg:relative lg:col-start-2 lg:row-start-1 lg:row-end-4 lg:block h-full">
+    <img src="https://cataas.com/cat" alt="" class="absolute inset-0 h-full w-full object-cover" />
+  </div>
+</div>
+
+{{ end }}
