+++ src/templates/contact.go.html
@@ -1,0 +1,10 @@
+{{ define "content" }}
+<div>
+    <h1>This is the contact page template</h1>
+    <div>
+        <p>Here is some content</p>
+
+        <img {{ .getImageProps "https://images.unsplash.com/photo-1682687981630-cefe9cd73072" }} />
+    </div>
+</div>
+{{ end }}
