+++ src/templates/index.go.html
@@ -1,13 +1,3 @@
-<!doctype html>
-<html lang="en" class="h-full dark">
-{{ template "head" . }}
-
-<body hx-ext="preload" class="flex min-h-screen flex-col justify-between">
-  <div>
-    {{ template "navbar" . }}
-
-    <main>{{ template "blocks" . }}</main>
-  </div>
-</body>
-
-</html>
+{{ define "content" }}
+{{ template "blocks" . }}
+{{ end}}
