+++ src/templates/layout.go.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+
+<html lang="en" class="dark h-full">
+{{ template "head" . }}
+
+<body hx-ext="preload" class="flex min-h-screen flex-col justify-between">
+    <div>
+        {{ template "navbar" . }}
+
+        <main>{{ template "content" . }}</main>
+
+        {{ template "footer" . }}
+    </div>
+</body>
+
+</html>
